fix: split dir unit/e2e
