+++ test/e2e/index.html
@@ -3,7 +3,7 @@
 
 <head>
   <meta charset="utf-8" />
-  <script src="./../umd/min.js"></script>
+  <script src="./../../umd/min.js"></script>
   <script>
     var holidays = holiday_jp_dayjs.between(new Date('2010-09-14T00:00+0900'), new Date('2010-09-21T00:00+0900'));
     console.log(holidays[0]['name'])
